Feature #16: Add babel, refactor code into modules, adjust common webpack config for babel
