+++ src/.eslintrc
@@ -1,13 +1,13 @@
 {
   "extends": "eslint:recommended",
-  "env": { 
+  "env": {
     "es6": true,
     "browser": true,
     "webextensions": true
   },
   "parserOptions": {
     "ecmaVersion": 6,
-    "sourceType": "script",
+    "sourceType": "module",
     "ecmaFeatures": {
       "impliedStrict": true
     }
