+++ src/package.json
@@ -10,6 +10,8 @@
     "@babel/core": "^7.2.2",
     "@babel/preset-env": "^7.3.1",
     "babel-loader": "^8.0.5",
+    "babel-minify-webpack-plugin": "^0.3.1",
+    "babel-preset-minify": "^0.5.0",
     "bestzip": "^2.1.2",
     "copy-webpack-plugin": "^4.6.0",
     "eslint": "^5.12.1",
